+++ src/main/java/net/acetheeldritchking/discerning_the_eldritch/utils/DTEServerConfig.java
@@ -77,16 +77,16 @@ public class DTEServerConfig
             .define("Sheath of The Razor Multiplier", 10);
 
     // Mourning Star
-    private static final ModConfigSpec.ConfigValue<Float> MOURNING_STAR_MORTAL_DAMAGE_MULTIPLIER = BUILDER
+    private static final ModConfigSpec.ConfigValue<Double> MOURNING_STAR_MORTAL_DAMAGE_MULTIPLIER = BUILDER
             .comment("Defines the multiplier for the Mourning Star's mortal damage decree")
             .comment("Default is [5.5F]")
-            .define("Mourning Star Damage Multiplier", 5.5F);
+            .define("Mourning Star Damage Multiplier", 5.5);
 
     // Exorcist's Aegis
-    private static final ModConfigSpec.ConfigValue<Float> EXORCISTS_AEGIS_FAIL_CHANCE = BUILDER
+    private static final ModConfigSpec.ConfigValue<Double> EXORCISTS_AEGIS_FAIL_CHANCE = BUILDER
             .comment("Defines the fail chance for casting an Eldritch spell with Exorcist's Aegis equipped")
             .comment("Default is [25.0F]")
-            .define("Exorcist's Aegis fail chance", 25.0F);
+            .define("Exorcist's Aegis fail chance", 25.0);
 
     // Boss Damage Caps //
     // Ascended One
@@ -154,9 +154,9 @@ public class DTEServerConfig
     public static int razorSheathCooldown;
     public static int razorSheathMultiplier;
     public static boolean gaolerWalkingScreenshake;
-    public static float mourningStarMortalDamageMultiplier;
+    public static double mourningStarMortalDamageMultiplier;
     public static boolean enableBloodCultistAggression;
-    public static float exorcistsAegisFailChance;
+    public static double exorcistsAegisFailChance;
 
 
     @SubscribeEvent
